feat(accordion): update mdx for accordion
update mdx and structure

#294
