+++ packages/bcgov-theme/stories/accordion/args.ts
@@ -1,0 +1,9 @@
+export const argTypes = {
+  onToggle: { action: 'toggling' },
+  defaultToggled: {
+    description: 'Set true to leave the panel open by default.',
+  },
+  title: {
+    description: 'The title of the accordion.',
+  },
+};
